+++ matlab/utils/CSF_st_fov.m
@@ -167,7 +167,7 @@ classdef CSF_st_fov
             sccsf = SCCSF_ConeContrastMat();
             LMS_d65 = xyz2lms2006( whitepoint( 'd65' ) );
             % Achromatic sensitivity, accounting for luminance, frequency and size
-            S_sp = sccsf.sensitivity_coldir( rho, L_bkg(:) * LMS_d65, 1, A_cm );
+            S_sp = sccsf.sensitivity_coldir( rho_cm, L_bkg(:) * LMS_d65, 1, A_cm );
             
             S = reshape( S_sp .* S_st, size(L_bkg) );