+++ ChangeLog.md
@@ -1,7 +1,8 @@
-# v1.01 - 11/11/2021
+# v1.01 - 15/11/2021
 - [Python] Optimized code for the Laplacian pyramid decomposition - faster predictions. There is a small (<0.02) difference in the predicted JODs.
 - [Matlab] Improved computation of eccentricity for wide-field-of-view displays
 - [Matlab] Added new option 'ignore_boundary_pixels'
+- [Matlab] Fixed the cortical magnificatioon bug in the CSF
 
 # v1.0 - 21/04/2021